move .proto and .pb.go files to be in the same directory
