+++ hegelc/main.go
@@ -8,7 +8,7 @@ import (
 	"io"
 	"log"
 
-	"github.com/tinkerbell/hegel/grpc/hegel"
+	"github.com/tinkerbell/hegel/grpc/protos/hegel"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
 )
